Removed debug error message on vim startup
Commented out the usage of a function to add aditional text to a status
line if the window displaying the status bar is not the current vim
window. This is still afeature that is in progress, and has not bee
completed. I have commented out the usage of the function to remove the
debug statment from vim startup.